+++ efx/src/tags/column.rs
@@ -3,9 +3,9 @@ use proc_macro2::TokenStream;
 use quote::{ToTokens, quote};
 
 use crate::render::render_nodes_as_stmts;
-use crate::tags::util::{attr_map, f32_opt};
 use crate::tags::{Block, TagAttributes};
 use efx_attrnames::AttrNames;
+use crate::utils::attr::*;
 
 pub struct Column {
     attributes: Attributes,
