+++ efx/src/tags/label.rs
@@ -1,10 +1,10 @@
-use crate::build_buffer_from_children;
-use crate::tags::util::{attr_map, bool_opt, color_tokens_opt, f32_opt};
 use crate::tags::{Tag, TagAttributes};
 use efx_attrnames::AttrNames;
 use efx_core::Element;
 use proc_macro2::TokenStream;
 use quote::{ToTokens, quote};
+use crate::utils::attr::*;
+use crate::utils::buffer::build_buffer_from_children;
 
 pub struct Label {
     attributes: Attributes,
