+++ efx/src/tags/row.rs
@@ -1,10 +1,10 @@
 use crate::render::render_nodes_as_stmts;
-use crate::tags::util::{attr_map, bool_or, f32_opt};
 use crate::tags::{Block, TagAttributes};
 use efx_attrnames::AttrNames;
 use efx_core::Element;
 use proc_macro2::TokenStream;
 use quote::{ToTokens, quote};
+use crate::utils::attr::*;
 
 pub struct Row {
     attributes: Attributes,
