+++ efx/src/tags/text_field.rs
@@ -1,10 +1,11 @@
-use crate::tags::util::*;
 use crate::tags::{Tag, TagAttributes};
 use efx_attrnames::AttrNames;
 use efx_core::Element;
 use proc_macro2::TokenStream;
 use quote::{ToTokens, quote};
 use syn::Expr;
+use crate::utils::attr::*;
+use crate::utils::expr::expr_req;
 
 pub struct TextField {
     attributes: Attributes,
